Moves All Enums to the Correct Dir
