+++ src/main/java/br/com/hideyoshi/auth/enums/Provider.java
@@ -1,4 +1,4 @@
-package br.com.hideyoshi.auth.entity;
+package br.com.hideyoshi.auth.enums;
 
 import lombok.Getter;
