+++ src/main/java/br/com/hideyoshi/auth/util/validator/password/PasswordValidator.java
@@ -1,6 +1,6 @@
 package br.com.hideyoshi.auth.util.validator.password;
 
-import br.com.hideyoshi.auth.entity.Provider;
+import br.com.hideyoshi.auth.enums.Provider;
 import lombok.RequiredArgsConstructor;
 
 import javax.validation.ConstraintValidator;
